Refactor directory structure, grouping airflow stuff
